backend/requirements.txt
aiohttp==3.9.4
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==2.0.12
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
croniter==2.0.3
cryptography==42.0.5
defusedxml==0.8.0rc2
distlib==0.3.7
Django==4.1.6
django-celery-beat==2.6.0
django-cors-headers==3.13.0
django-extensions==3.2.3
django-filter==23.5
django-sslserver==0.22
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-nested-routers==0.93.5
drf-yasg==1.21.7
et-xmlfile==1.1.0
filelock==3.12.2
frozenlist==1.4.1
gunicorn==22.0.0
idna==3.7
inflection==0.5.1
kombu==5.3.7
Markdown==3.6
MarkupSafe==2.1.5
multidict==6.0.5
numpy==1.26.4
oauthlib==3.2.2
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pipenv==2023.8.19
platformdirs==3.10.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-crontab==3.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.0
requests==2.27.1
requests-oauthlib==2.0.0
rest-social-auth==8.0.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.18
vine==5.1.0
virtualenv==20.24.3
wcwidth==0.2.13
Werkzeug==3.0.3
yarl==1.9.4
